refactor(onboarding): file structure change
