+++ src/modules/Common/Authentication/pages/Onboarding/SignIn/signin.tsx
@@ -1,7 +1,7 @@
 import styles from "./SignIn.module.css";
 import google from "../../../assets/google.png";
-import OnboardingTemplate from "../OnboardingTeamplate/OnboardingTemplate";
-import OnboardingHeader from "../../../components/Head/OnboardingHeader";
+import OnboardingTemplate from "../../../components/OnboardingTeamplate/OnboardingTemplate";
+import OnboardingHeader from "../../../components/OnboardingHeader/OnboardingHeader";
 
 export default function SignIn() {
     return (
